+++ crates/ryo3-fspath/src/lib.rs
@@ -1,6 +1,7 @@
 //! `FsPath` python module
 use pyo3::prelude::*;
 mod fspath;
+mod pathlib;
 pub use fspath::PyFsPath;
 
 pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {